test(Amount): Amount.parseAmount 테스트를 추가하라
Closes
